review: DeleteIfExists reuses IsNotFound; int64 byte-count compare; doc Metadata aliasing
Addresses external review + Copilot on #11.
- DeleteIfExists now calls IsNotFound (single source of truth for not-found).
- hash test compares byte count as int64 (no int conversion / overflow risk).
- TextArtifacts/SingleTextArtifact docs note the returned artifacts are shallow
  copies sharing Metadata (use content.Artifact.Clone for independence).

+++ content/hash_test.go
@@ -50,7 +50,7 @@ func TestSHA256HexReaderAgreesWithBytes(t *testing.T) {
 	if got != want {
 		t.Fatalf("reader hash %q != string hash %q", got, want)
 	}
-	if int(n) != len(s) {
+	if n != int64(len(s)) {
 		t.Fatalf("n = %d, want %d", n, len(s))
 	}
 }
